Jdk21

+++ src/main/java/com/baiyx/wfwbitest/Controller/Service/ServiceImpl/Dao/ProjBaseDao.java
@@ -19,12 +19,12 @@ public interface ProjBaseDao {
 
     // 查询projbase表总记录条数
     @DS("slave_1")
-    int CountProjbase();
+    long CountProjbase();
 
     Long selectCarArchivesList_COUNT();
 
     // 循环查询,每次查询固定的条数
-     @DS("slave_1")
+    @DS("slave_1")
     List<Projbase> readProjbase(Map map);
 
     // 写入数据
